Fix c-style casts

+++ onnx/defs/math/defs.cc
@@ -2059,9 +2059,9 @@ static bool BuildContextDependentFunctionBody(
       reduction_attr_proto != nullptr && reduction_attr_proto->has_s() ? reduction_attr_proto->s() : "mean";
 
   FunctionBuilder builder(functionProto);
-  builder.Const1D("const_zero", int64_t(0))
-      .Const1D("const_one", int64_t(1))
-      .Const1D("axes", int64_t(1))
+  builder.Const1D("const_zero", 0)
+      .Const1D("const_one", 1)
+      .Const1D("axes", 1)
       .Add("expanded_target = Unsqueeze (target, axes)");
 
   if (ctx.getAttribute("ignore_index") == nullptr) {
